+++ CHANGELOG.md
@@ -5,6 +5,9 @@ This changelog references the relevant changes done between versions.
 To get the diff for a specific change, go to https://github.com/LIN3S/FrontFoundation/commit/XXX where XXX is the change hash 
 To get the diff between two versions, go to https://github.com/LIN3S/FrontFoundation/compare/v0.5.0...v0.6.0
 
+* 0.11.0
+    * Added parsleyjs validation both to the `FormSelect` react and vanilla atoms implementations.
+    * Added required prop to the `FormGroupInput` react component.
 * 0.10.0
     * Added full keyboard integration to the `FormSelect` vanilla js atom.
     * Fixed tabindex prop name in the `FormSelect` react atom.